django/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-allauth==0.64.0
django-oauth-toolkit==1.7.1
django-guardian==2.4.0
django-anymail==10.2
whitenoise==6.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
boto3==1.34.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
